requirements.txt
dash>=2.18.0
dash-bootstrap-components>=1.6.0
httpx>=0.27.0
pandas>=2.2.0
plotly>=5.24.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
claude-agent-sdk>=0.1.30
anyio>=4.0.0
